fix all unsigned 'foreign key' columns in db, refactor handlers + db entities + models out of main.rs
