+++ Cargo.toml
@@ -21,6 +21,7 @@ serde_json = "1.0.97"
 sha2 = "0.10"
 sqlx = { version = "0.6", features = [ "runtime-actix-native-tls", "mysql", "chrono", "uuid" ] }
 tera = "1"
+thiserror = "1.0.40"
 
 [dependencies.uuid]
 version = "1.3.4"
