+++ schema/comments.sql
@@ -1,10 +1,10 @@
 CREATE TABLE `comments` (
     `id` bigint unsigned NOT NULL AUTO_INCREMENT,
     `public_id` binary(16) NOT NULL,
-    `author` bigint NOT NULL,
-    `post_id` bigint NOT NULL,
-    `parent_id` bigint NULL,
-    `content_id` bigint NOT NULL,
+    `author` bigint unsigned NOT NULL,
+    `post_id` bigint unsigned NOT NULL,
+    `parent_id` bigint unsigned NULL,
+    `content_id` bigint unsigned NOT NULL,
     `created` datetime NOT NULL,
     `updated` datetime NOT NULL,
