+++ schema/post_content_audit.sql
@@ -1,7 +1,7 @@
 CREATE TABLE `post_content_audit` (
     `id` bigint unsigned NOT NULL AUTO_INCREMENT,
-    `post_id` bigint NOT NULL,
-    `content_id` bigint NOT NULL,
+    `post_id` bigint unsigned NOT NULL,
+    `content_id` bigint unsigned NOT NULL,
     `created` datetime NOT NULL,
     
     PRIMARY KEY (`id`),
