+++ schema/posts.sql
@@ -1,10 +1,10 @@
 CREATE TABLE `posts` (
     `id` bigint unsigned NOT NULL AUTO_INCREMENT,
     `public_id` binary(16) NOT NULL,
-    `author` bigint NOT NULL,
+    `author` bigint unsigned NOT NULL,
     `title` varchar(512) NOT NULL,
     `link` varchar(1024) NULL,
-    `content_id` bigint NULL,
+    `content_id` bigint unsigned NULL,
     `created` datetime NOT NULL,
     `updated` datetime NOT NULL,
