+++ schema/users.sql
@@ -2,7 +2,7 @@ CREATE TABLE `users` (
     `id` bigint unsigned NOT NULL AUTO_INCREMENT,
     `public_id` binary(16) NOT NULL,
     `name` varchar(64) NOT NULL,
-    `email_id` bigint NOT NULL,
+    `email_id` bigint unsigned NOT NULL,
     `password` varchar(1024) NOT NULL, -- hash:salt:hash_func
     `is_deleted` boolean NOT NULL,
     `created` datetime NOT NULL,
