+++ src/entities/mod.rs
@@ -1,0 +1,2 @@
+pub mod email;
+pub mod user;
