+++ src/routes/mod.rs
@@ -1,0 +1,7 @@
+mod models;
+
+pub mod health;
+pub mod home;
+pub mod login;
+pub mod signup;
+pub mod submit;
