+++ src/routes/models/mod.rs
@@ -1,0 +1,6 @@
+mod post;
+mod user;
+
+pub use post::Post;
+pub use user::translate_user;
+pub use user::User;
