+++ src/routes/models/post.rs
@@ -1,0 +1,8 @@
+use serde::Serialize;
+
+#[derive(Serialize)]
+pub struct Post {
+    pub author: String,
+    pub title: String,
+    pub body: String,
+}
