+++ src/routes/signup/mod.rs
@@ -1,0 +1,2 @@
+pub mod get;
+pub mod post;
