+++ src/routes/submit/mod.rs
@@ -1,0 +1,2 @@
+pub mod get;
+pub mod post;
